move entityTool to own file to avoid circular dependency
